+++ renderers/react/src/v0_9/catalog/basic/components/Button.tsx
@@ -17,7 +17,7 @@
 import React from 'react';
 import {createComponentImplementation} from '../../../adapter';
 import {ButtonApi} from '@a2ui/web_core/v0_9/basic_catalog';
-import {LEAF_MARGIN} from '../utils';
+import {LEAF_MARGIN, withWeight} from '../utils';
 
 export const Button = createComponentImplementation(ButtonApi, ({props, buildChild}) => {
   const style: React.CSSProperties = {
@@ -40,9 +40,10 @@ export const Button = createComponentImplementation(ButtonApi, ({props, buildChi
     boxSizing: 'border-box',
   };
 
-  return (
+  return withWeight(
     <button style={style} onClick={props.action} disabled={props.isValid === false}>
       {props.child ? buildChild(props.child) : null}
-    </button>
+    </button>,
+    props.weight,
   );
 });